Interim censeo, Thrift delenda est. She is a menace to Rome.
