+++ index.js
@@ -100,10 +100,7 @@ function CassandraAdapter()
 	this.attachfamily = null;
 	this.tables       = {};
 	this.connection   = null;
-	this.withKeyspace = null;
-	this.keyspace     = null;
 	this.columnFamily = null;
-	this.attachments  = null;
 }
 
 CassandraAdapter.prototype.configure = function(options, modelfunc)
@@ -113,7 +110,6 @@ CassandraAdapter.prototype.configure = function(options, modelfunc)
 	_.assign(this.options, options);
 	this.constructor = modelfunc;
 	this.family = modelfunc.prototype.plural;
-	this.attachfamily = this.family + '_attachments';
 
 	// If you hand us a pool, we presume you have added an error listener.
 	if (options.connection)
@@ -133,35 +129,11 @@ CassandraAdapter.prototype.configure = function(options, modelfunc)
 			throw(err);
 		});
 	}
-
-	this.withKeyspace = this.connection.connect().then(function()
-	{
-		return self.connection.assignKeyspace(self.options.keyspace);
-	}).then(function(ks)
-	{
-		if (Array.isArray(ks))
-		{
-			var promise = _.find(ks, { 'state': 'fulfilled' });
-
-			if (!promise)
-			{
-				var invalidError = new Error('Missing or invalid response.');
-				_.assign(invalidError, { 'keyspace': self.options.keyspace, 'response': ks });
-				throw invalidError;
-			}
-
-			self.keyspace = promise.value;
-		}
-		else
-			self.keyspace = ks;
-
-		return self.keyspace;
-	});
 };
 
 CassandraAdapter.prototype.getAttachmentTable = function()
 {
-	return this.keyspace.getTableAs(this.attachfamily, 'attachments');
+	return P();
 };
 
 var typeToValidator =
@@ -219,6 +191,7 @@ CassandraAdapter.prototype.createModelTable = function()
 	query += ', PRIMARY KEY (' + throwaway.keyfield + '))';
 
 	return self.connection.cql(query)
+	.then(function() { return 'OK'; })
 	.fail(function(error)
 	{
 		if (/^Cannot add already existing/i.test(error.why))
@@ -228,48 +201,57 @@ CassandraAdapter.prototype.createModelTable = function()
 	});
 };
 
-CassandraAdapter.prototype.createAttachmentsTable = function()
+CassandraAdapter.prototype.createKeyspace = function()
 {
 	var self = this;
 
-	return this.keyspace.createTableAs(this.attachfamily, 'attachments',
+	var query = new Query("CREATE KEYSPACE {keyspace} WITH REPLICATION = {replication};")
+	.params(
 	{
-		description: 'polyclay ' + this.constructor.prototype.singular + ' attachments',
-		columns:
-		[
-			{ name: 'name',         validation_class: 'UTF8Type'  },
-			{ name: 'content_type', validation_class: 'UTF8Type'  },
-			{ name: 'data',         validation_class: 'AsciiType' }
-		]
-	})
-	.then(function(table)
+		'keyspace': self.options.keyspace,
+		replication: {'class' : 'SimpleStrategy', 'replication_factor': 3 }
+	}).types({ replication: 'map<text, text>'});
+
+	return query.execute(self.connection)
+	.then(function()
 	{
-		self.attachments = table;
-		return table;
+		return 'OK';
+	}).fail(function(error)
+	{
+		if (/^Cannot add existing/i.test(error.why))
+			return 'OK';
+
+		throw error;
 	});
 };
 
 CassandraAdapter.prototype.provision = function(callback)
 {
-	if (this.options.noprovision)
-		return callback(null, 'OK');
-
 	var self = this;
 
-	return this.withKeyspace
+	self.connection.connect()
 	.then(function()
 	{
-		return P.all(
-		[
-			self.createModelTable(),
-			self.createAttachmentsTable()
-		]);
+		if (self.options.noprovision)
+			return callback(null, 'OK');
+
+		return self.createKeyspace();
 	})
 	.then(function()
 	{
-		callback(null, 'OK');
-	}, callback)
-	.done();
+		var query = new Query('use {keyspace};').params({'keyspace': self.options.keyspace});
+		return query.execute(self.connection);
+	})
+	.then(function() { return self.createModelTable(); })
+	.then(function(response)
+	{
+		callback(null, response);
+	})
+	.fail(function(error)
+	{
+		console.log(error);
+		callback(error);
+	}).done();
 };
 
 CassandraAdapter.prototype.shutdown = function() {};
@@ -302,10 +284,7 @@ CassandraAdapter.prototype.save = function(obj, json, callback)
 	q2 = q2.slice(0, q2.length - 2);
 	var query = new Query(q1 + q2 + ')').types(types).params(params);
 
-	return this.withKeyspace
-	.then(function() { return query.execute(self.connection); })
-	.then(function() { return self.createAttachmentsTable(); })
-	.then(function() { return self.saveAttachments(obj.key, json._attachments); })
+	return query.execute(self.connection)
 	.then(function(resp) { callback(null, 'OK');}, callback)
 	.done();
 };
@@ -338,8 +317,7 @@ CassandraAdapter.prototype.merge = function(key, properties, callback)
 	query = query.slice(0, query.length - 2);
 	query += ' WHERE {keyfield} = {key}';
 
-	return this.withKeyspace
-	.then(function() { return new Query(query).params(params).types(types).execute(self.connection); })
+	return new Query(query).params(params).types(types).execute(self.connection)
 	.then(function()
 	{
 		callback(null, 'OK');
@@ -349,23 +327,7 @@ CassandraAdapter.prototype.merge = function(key, properties, callback)
 
 CassandraAdapter.prototype.saveAttachment = function(obj, attachment, callback)
 {
-	var self = this;
-	var key = makeAttachKey(obj.key, attachment.name);
-
-	var values =
-	{
-		name:         attachment.name,
-		content_type: attachment.content_type,
-		data:         new Buffer(attachment.body).toString('base64')
-	};
-
-	return this.getAttachmentTable()
-	.then(function() { return self.attachments.insert(key, values); })
-	.then(function(res)
-	{
-		callback(null, 'OK');
-	}, callback)
-	.done();
+	callback(null, 'OK');
 };
 
 CassandraAdapter.prototype.saveAttachments = function(key, attachments)
@@ -420,7 +382,7 @@ CassandraAdapter.prototype.get = function(key, callback)
 		'key': typeToValidator[throwaway.propertyType(throwaway.keyfield)]
 	});
 
-	this.withKeyspace.then(function() { return query.execute(self.connection); })
+	query.execute(self.connection)
 	.then(function(rows)
 	{
 		rows.forEach(function(row)
@@ -468,7 +430,7 @@ CassandraAdapter.prototype.getBatch = function(keylist, callback)
 		'keyfield': self.constructor.prototype.keyfield
 	}).params(params).types(types);
 
-	this.withKeyspace.then(function() { return query.execute(self.connection); })
+	return query.execute(self.connection)
 	.then(function(rows)
 	{
 		rows.forEach(function(row)
@@ -492,9 +454,10 @@ CassandraAdapter.prototype.all = function(callback)
 	var results = [];
 	var self = this;
 
-	this.withKeyspace.then(function() { return new Query('SELECT * from {keyspace}.{family}')
-	.params({ 'family': self.family, 'keyspace': self.options.keyspace })
-	.execute(self.connection); })
+	var q = new Query('SELECT * from {keyspace}.{family}');
+
+	q.params({ 'family': self.family, 'keyspace': self.options.keyspace })
+	.execute(self.connection)
 	.then(function(rows)
 	{
 		rows.forEach(function(row)
@@ -525,7 +488,7 @@ CassandraAdapter.prototype.attachment = function(key, name, callback)
 	var cassKey = makeAttachKey(key, name);
 	var query = 'SELECT * from ' + self.attachfamily + ' WHERE key = ?';
 
-	this.withKeyspace.then(function() { return self.connection.cql(query, [cassKey]); })
+	self.connection.cql(query, [cassKey])
 	.then(function(rows)
 	{
 		if (rows.length === 0)
@@ -574,7 +537,7 @@ CassandraAdapter.prototype.remove = function(obj, callback)
 		'key': typeToValidator[obj.propertyType(obj.keyfield)]
 	});
 
-	this.withKeyspace.then(function() { return query.execute(self.connection); })
+	query.execute(self.connection)
 	.then(function(reply)
 	{
 		return self.removeAllAttachments(obj.key);
