+++ package.json
@@ -3,6 +3,7 @@
 	"version":        "1.0.7",
 	"description":    "cassandra persistence adapter for polyclay, the schema-enforcing document mapper",
 	"author":         "C J Silverio <ceejceej@gmail.com>",
+	"contributors":   [ "C J Silverio <ceejceej@gmail.com>", "Kit Cambridge <kitcambridge@me.com>" ],
 	"license":        "MIT",
 	"readmeFilename": "README.md",
 	"main":           "index.js",
